# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "CrestedIbis目前是一个基于GB28181标准实现的音视频云平台，负责实现GB28181信令和设备管理，未来将会是一个支持物联网设备接入，算法训练和部署的综合物联网平台。",
        "title": "CrestedIbis",
        "contact": {
            "name": "北溪入江流(focus1024)",
            "url": "http://focus1024.com(https://github.com/focus1024-wind/CrestedIbis)",
            "email": "focus1024@foxmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "paths": {
        "/ipc/device/channels": {
            "get": {
                "description": "查询GB28181 设备对应通道信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IPC设备 /ipc/device"
                ],
                "summary": "获取设备通道信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "访问token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "访问token",
                        "name": "access_token",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "device_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询数据失败",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.HttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/ipc/device/devices": {
            "get": {
                "description": "分页查询GB28181 IpcDevice设备",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IPC设备 /ipc/device"
                ],
                "summary": "分页查询IpcDevice设备",
                "parameters": [
                    {
                        "type": "string",
                        "description": "访问token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "访问token",
                        "name": "access_token",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "分页查询页码，默认值: 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页查询数量，默认值: 15",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询数据失败",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.HttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/ipc/device/upload_image": {
            "post": {
                "description": "GB28181图像抓拍，图片上传接口",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IPC设备 /ipc/device"
                ],
                "summary": "IPC图像上传",
                "parameters": [
                    {
                        "type": "string",
                        "description": "访问token",
                        "name": "access_token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "上传图片",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "上传图片失败",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.HttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/system/user/login": {
            "post": {
                "description": "用户登录并生成用户登录日志信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理 /system/user"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "用户登录信息，密码采用加盐加密",
                        "name": "SysLoginUser",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.SysUserLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功，响应JWT",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.HttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "登录失败，响应失败信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.HttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/system/user/register": {
            "post": {
                "description": "注册用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理 /system/user"
                ],
                "summary": "注册用户",
                "parameters": [
                    {
                        "description": "用户注册信息，密码采用加盐加密",
                        "name": "SysUser",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.SysUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注册成功",
                        "schema": {
                            "$ref": "#/definitions/model.HttpResponse"
                        }
                    },
                    "500": {
                        "description": "注册失败，响应失败信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.HttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "ipc_device.IpcChannel": {
            "type": "object",
            "properties": {
                "created_time": {
                    "type": "string"
                },
                "device_id": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "manufacturer": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                },
                "state": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "updated_time": {
                    "type": "string"
                }
            }
        },
        "ipc_device.IpcDevice": {
            "type": "object",
            "properties": {
                "created_time": {
                    "type": "string"
                },
                "device_addr": {
                    "type": "string"
                },
                "device_id": {
                    "type": "string"
                },
                "from_address": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ipcChannels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ipc_device.IpcChannel"
                    }
                },
                "manufacturer": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "register_time": {
                    "type": "string"
                },
                "sn": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "updated_time": {
                    "type": "string"
                }
            }
        },
        "ipc_device.IpcDevicePage": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ipc_device.IpcDevice"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "model.HttpResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "user.RoleGroup": {
            "type": "object",
            "properties": {
                "created_time": {
                    "type": "string"
                },
                "role_id": {
                    "type": "integer"
                },
                "role_name": {
                    "type": "string"
                },
                "updated_time": {
                    "type": "string"
                },
                "user": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user.SysUser"
                    }
                }
            }
        },
        "user.SysUser": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "created_time": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "example": "CrestedIbis"
                },
                "phone": {
                    "type": "string"
                },
                "roleGroups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user.RoleGroup"
                    }
                },
                "sex": {
                    "type": "integer"
                },
                "updated_time": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
        "user.SysUserLogin": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "CrestedIbis"
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        }
    }
}